+++ src/main/scala/wdlTools/cli/Upgrade.scala
@@ -1,11 +1,36 @@
 package wdlTools.cli
 
-import wdlTools.generators.code.Upgrader
+import wdlTools.generators.code.{Upgrader, Utils => GeneratorUtils}
 import dx.util.{FileNode, FileSourceResolver}
+import wdlTools.syntax.{Parsers, SyntaxException, WdlVersion}
+import wdlTools.types.{TypeCheckingRegime, TypeException, TypeInfer}
 
 import scala.language.reflectiveCalls
 
 case class Upgrade(conf: WdlToolsConf) extends Command {
+  def parseAndCheck(source: FileNode,
+                    wdlVersion: WdlVersion,
+                    followImports: Boolean,
+                    message: String): Unit = {
+    val errorHandler = new CliTypeErrorHandler()
+    val parsers = Parsers(followImports)
+    val parser = parsers.getParser(wdlVersion)
+    val checker = TypeInfer(TypeCheckingRegime.Moderate, errorHandler = Some(errorHandler))
+    try {
+      checker.apply(parser.parseDocument(source))
+    } catch {
+      case e: SyntaxException =>
+        throw new Exception(s"Failed to parse ${source}; ${message}", e)
+      case e: TypeException =>
+        throw new Exception(s"Failed to type-check ${source}; ${message}", e)
+    }
+    if (errorHandler.hasTypeErrors) {
+      throw new Exception(
+          s"Failed to type-check ${source}; ${message}:\n${errorHandler.toString}"
+      )
+    }
+  }
+
   override def apply(): Unit = {
     val docSource = FileSourceResolver.get.resolve(conf.upgrade.uri())
     val srcVersion = conf.upgrade.srcVersion.toOption
@@ -16,7 +41,7 @@ case class Upgrade(conf: WdlToolsConf) extends Command {
     val upgrader = Upgrader(followImports)
     // write out upgraded versions
     val documents = upgrader.upgrade(docSource, srcVersion, destVersion)
-    val upgraded = writeDocuments(documents, outputDir, overwrite)
+    val upgraded = GeneratorUtils.writeDocuments(documents, outputDir, overwrite)
     if (conf.upgrade.check() && upgraded.nonEmpty) {
       upgraded.head._1 match {
         case fn: FileNode =>